Fixed prisma imports after prisma/index.ts file rename / location change
