+++ pages/api/gifts/[uuid].ts
@@ -1,6 +1,6 @@
 import { Gift } from '~/shared/types';
 import { NextApiRequest, NextApiResponse } from 'next';
-import prisma from '~/utils/prismaClient';
+import prisma from '~/prisma';
 import { errorFound } from '.';
 
 type HandlerParams = {
